moved torch's adaptive avg pooling functions to pooling_functions.py
